refactor: 숫자패키지로 이동
