+++ src/main/java/baseball/domain/digits/Digit.java
@@ -1,4 +1,4 @@
-package baseball.domain;
+package baseball.domain.digits;
 
 import java.util.Objects;
