+++ src/main/java/baseball/domain/digits/IndexedDigitIterator.java
@@ -1,4 +1,4 @@
-package baseball.domain;
+package baseball.domain.digits;
 
 import java.util.Iterator;
 import java.util.List;
