+++ src/test/java/baseball/domain/digits/DigitTest.java
@@ -1,4 +1,4 @@
-package baseball.domain;
+package baseball.domain.digits;
 
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.assertj.core.api.Assertions.assertThatIllegalArgumentException;
