+++ src/test/java/baseball/domain/digits/DigitsUtil.java
@@ -1,4 +1,4 @@
-package baseball.domain;
+package baseball.domain.digits;
 
 import java.util.ArrayList;
 import java.util.List;
